DreamBooth_Fine_Tuning_Text_to_Image_Diffusion_Models_for_Subject_Driven_Generation/requirements.txt
diffusers==0.30.3
matplotlib==3.5.1
Pillow==9.0.1
torch==2.7.0
torchvision==0.22.0
tqdm==4.67.1
transformers==4.46.3